+++ src/supply_pile.rs
@@ -1,11 +1,11 @@
 use crate::card::Card;
 pub struct SupplyPile<'a> {
     pub card: &'a Card,
-    pub count: u8,
+    pub count: u16,
 }
 
 impl<'a> SupplyPile<'a> {
-    pub fn new(card: &'a Card, count: u8) -> Self {
+    pub fn new(card: &'a Card, count: u16) -> Self {
         SupplyPile { card, count }
     }
 }
